requirements.txt
opencv-python
numpy
face-recognition
Pillow
